requirements.txt
flask>=2.3.0
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
python-magic-bin>=0.4.14
typer>=0.9.0
rich>=13.0.0
loguru>=0.7.0
waitress>=2.1.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.0.0
psutil>=5.9.0 